Withdraw: switch co-sign → custodial-push (no wallet signing)
Co-sign требовал подпись игроком заранее собранной транзакции — Phantom показывал
красное «dApp may be malicious» на новом домене petix.fun, отпугивая пользователей.

Custodial-push: сервер сам отправляет токены (treasury = fee payer + источник,
полностью подписывает и шлёт). Игрок ничего не подписывает → нет промпта/предупреждения
кошелька. Цена: газ + аренду ATA (~0.002 SOL/новый кошелёк) платим мы → казне нужен SOL.

- api/_lib/withdraw.js: buildWithdrawTransaction (co-sign) → sendWithdrawFromTreasury
  (custodial: feePayer=treasury, ATA payer=treasury, sign+sendRawTransaction).
- Единый POST /api/withdraw/request: резерв-дебет Points → отправка → поллинг
  подтверждения → возврат Points при сбое отправки/tx. Удалены prepare/confirm/cancel.
- Фронт: модалка теперь делает один вызов request; убраны web3.js-загрузчик, резолв
  провайдера, signAndSendTransaction (браузеру Solana-SDK больше не нужен).

Проверено реальным custodial-выводом на mainnet (Token-2022 тест-токен): tx confirmed,
200 токенов ушли с казны, Points списались; резерв/возврат при пустой казне работает.
155/155 тестов зелёные.
